Emacs mu4e: fix for timezone problem
The timezone of the citation line was always GMT which could be confusing for
some recipients. An comment on
https://groups.google.com/forum/#!topic/mu-discuss/JlDSKsm9X5Y provided a fix
for this problem.

The switch from %H:%M:%S to %Z is purly code-cosmetic and doesn't change the
output.

The setting of the timezone with setenv is not part of the comment mentioned
above.